publish.sh
The publish.sh script does the following:

1. runs "lerna publish --no-git --no-npm"
This ask about what version to set, and it depends
on the semantic nature of the update (breaking/
non-breaking).

2. Invoke "npm pack" in all non-private modules
This produces an npm tarball for each module.

3. Zip all these tarballs into 

After build is successful, go to all non-private
modules and run "npm pack", which bundles the 
module into a tarball ready to be published into
an npm repository.

Update README and remove LICENSE file for now since
we are starting with a private beta.