Initial external release: jsii-java-runtime (#9)
* Initial external release: jsii-java-runtime

jsii runtime client for Java - allows generated jsii proxy classes to interact with jsii-runtime.

This is a maven package modeled as an npm package. It requires JDK8 and Maven installed in the build environment.

To build:

    npm run prepare

To pack:

    npm pack

Will result in an npm tarball that contains the resulting maven .jar.