Beta release bundle (#18)
* Java: remove hard-coded jsii-runtime version in handshake

When handshaking the java jsii client and jsii-runtime, we ensure that
their versions are compatible. We now extract the version package.json
during code generation and generate a Java constant with this value.

This removes the hard-coding of version value into the Java client.

* Beta release bundle

If build is successful, produces a zip under dist
which contains all the jsii tarballs.

This is how we package for beta.

* Add version to buildspec

* Fix conditions

* Generate version of expected calc assemblies based on repo version

Since in CI/CD we are creating a version that contains the commit hash
we need the expected assembly versions to correspond to it. This will also
ensure that when we bump MV, these expectations are automatically updated.

* Extract common build tools into a separate private package

align-jsii-versions will update the top-level and per-dependency `version` fields
for the specified assemblies. This can be used to align test expectations with
the current repo's version.

diff-test compares to assemblies (after aligning versions) and fails if they
are different, along with instructions on how to fix.

* Reduce loglevel

* Fix dep version update

* Make build-tools private