Stop inlinding dependencies' assemblies entirely
This produced extremely large assemblies, because the type models for
low-level libraries would be repeated an inordinate amount of times in
each and every assembly.

This commit makes JSII only track the module name and version of the
dependencies' assemblies, and adds a new `externalTypes` attribute to
the assembly to model type information of external (aka from
dependencies) types that are used in the signature of the local
assembly's exported signatures.

This commit also stops producing the deprecated `jsii.js` source bundle,
which is no longer necessary and not used by any runtime, and renames
the `jsii.json` file to `assembly.jsii`, for consistency with what the
java runtime bundles.

Fixes: awslabs/aws-cdk#56