Correct temporary directory creation
`fs.mkdtemp` does not append a path separator when creating the temporary
directory, and `os.tmpdir()` is not path-separator-terminated, so one needs
to manually add a path separator in order to correctly obtain a temporary
directory. This bug causes JSII builds to fail on systems where the user
is not authorized to create directories in the parent of `os.tmpdir()`.