java-client: embed jsii-runtime as a reesource
To remove the requirement for installing jsii-runtime
in the system, this change embeds a webpack'd version
of the jsii-runtime program as a Java resource and
loads it by default (if `JSII_RUNTIME` is unset).

This should dramatically improve the reliability and
compatibility of the client/runtime as they are now
self-contained and coupled together.