Bundle jsii modules as npm tarballs
This change modifies the way we bundle jsii modules. Instead of webpacking them
and storing the code inside the assembly spec, we now simply use "npm pack" to
produce an npm tarball (.tgz) and include that in the generated library.

The kernel now simply accepts the path to the tarball (as oppose to the entire
code passed through STDIN), and untars it into a working directory (under node_modules).
This effectively allows dependencies to `require(it)`.

This change also dramatically improves load-time performance (again, we are not sending
a big JSON file over the wire, parse it, and then evaluate it into the VM).

Bundled dependencies are now simple npm bundled dependencies. If a module uses the old-
style configuration, we fail with instructions.

NOTE: we currently do not support multiple versions of the same module loaded together
into the same kernel space. This is a major and unacceptable limitation for production
environments, which we must address, or otherwise, people will constantly hit these
restrictions as jsii software stacks evolve.

BREAKING CHANGE.