.NET: bundling of jsii-runtime and jsii tarball modules
Bundle jsii tarball modules into generated .NET packages and also
bundle the jsii-runtime program as an embedded resource into the
dotnet runtime library.

This removes the dependency on jsii-runtime, and leaves only node.js
as an external dependency. It also improves load-time performance since
the tarball is only passed as a path to the runtime/kernel and not
the entire .js code.