Bundle jsii modules as npm tarballs and embed jsii-runtime in .NET (#52)
This change modifies the way we bundle jsii modules. Instead of webpacking them
and storing the code inside the assembly spec, we now simply use "npm pack" to
produce an npm tarball (.tgz) and include that in the generated library.

The kernel now simply accepts the path to the tarball (as oppose to the entire
code passed through STDIN), and untars it into a working directory (under node_modules).
This effectively allows dependencies to `require(it)`.

This change also dramatically improves load-time performance (again, we are not sending
a big JSON file over the wire, parse it, and then evaluate it into the VM).

Bundled dependencies are now simple npm bundled dependencies. If a module uses the old-
style configuration, we fail with instructions.

NOTE: we currently do not support multiple versions of the same module loaded together
into the same kernel space. This is a major and unacceptable limitation for production
environments, which we must address, or otherwise, people will constantly hit these
restrictions as jsii software stacks evolve.

BREAKING CHANGE.

* CR fixes

* .NET: bundling of jsii-runtime and jsii tarball modules

Bundle jsii tarball modules into generated .NET packages and also
bundle the jsii-runtime program as an embedded resource into the
dotnet runtime library.

This removes the dependency on jsii-runtime, and leaves only node.js
as an external dependency. It also improves load-time performance since
the tarball is only passed as a path to the runtime/kernel and not
the entire .js code.

* Bring in a few fixes from dotnet-tarballs

* Initial set of CR comments

* Fix "update-version.sh"

`set -euo` fails for unbound variables

* Add CodeBuild build badge

* Add missing tgz to pacmak tests

* Ignore +build when setting jsii version

* update versions

* Require native type

* Do not include commit sha in version

Since we have version numbers in test expectations
every build will get a different version number and that
invariably causes tests to fail.

* Improvements to pacmak tests

1. Add calc-lib
2. Create expected tarball on-the-fly to avoid binary diffs resulting from different platforms

* Take only first 7 chars from commit sha

* Add node_modules/.bin to PATH so we can execute lerna

* Fix a couple of tests