Allow incremental build of documentation (#71)
Building documentation is rather slow, and when working on documentation
improvements this can be a real pain. Sphinx is however smart enough to
not fully re-process files that haven't changed since the last
execution.

This change adds a fingerprint to the generated `.rst` files that allows
`jsii-pacmak` to determine if they need be re-generated or not, such
that files for which the JSII assembly hasn't changed will not be
re-written, making documentation iteration much faster.