Fix representation of enum types in the assemblies
`enum` types are special union types in the TypeScript world. When an
`enum` is used in another union type (for example: an optional `enum`),
`tsc` represents it as the union of `undefined` and all the possible
`enum` values. Getting the *apparent* type of an `enum` value yields
it's *value type* (`string` or `number`), and not it's *base type*
(the `enum` type). Fortunately, the `typeChecker` provides an API to get
the base type of a litteral type (such as an `enum` value), which
correctly yields the `enum` type as desired.

This commit changes how the compiler resolves `enum`-like types for
inclusion in the assembly, and adds a couple of entries to `jsii-calc`
that advertised the incorrect behavior mentioned in #68, and that now
are correctly represented.

Fixes #68