Model `packageNames` on the JSII assemblies
When packaging artifacts for various package managers (`npm`, `mvn`,...)
it is often required to determine the name of the packages in order to
determine the relevant list of dependencies (for example, `mvn` requires
a `groupId` and `artifactId`). This information is also helpful when
determining how to publish packages to the package managers.

The new `packageNames` attribute is surfaced in the JSII assemblies, and
can be configured in a `package.json`'s `jsii` directive.

Smart defaults are provided for:
* `mvn`: the name is modeled as a string, in the way that is usually
  formatted in `mvn` commands: `<groupId>:<packageId>`. The elements are
  inferred from the `java` package name configured in `package.json`
  (at `jsii.names.java`):
  + `groupId` is the package name, excluding the last element
  + `artifactId` is the last element of the package name.
* `npm`: the name is the assembly's name, since this is already the
  `npm` name for the package. This key might look redundant, but it is
  included for consistency, and to avoid special-casing in tools that
  work with `npm`.