BREAKING: Model package names in JSII assemblies, and fingerprint assemblies (#85)
When packaging artifacts for various package managers (`npm`, `mvn`,...)
it is often required to determine the name of the packages in order to
determine the relevant list of dependencies (for example, `mvn` requires
a `groupId` and `artifactId`). This information is also helpful when
determining how to publish packages to the package managers.

The new `targets` attribute of the JSII assemblies replaces the `names`
attribute, and the information is no longer repeated in the `nativenames`
attribute. `targets` can be configured in the `jsii.targets` key of `package.json`.

The `dependencies` attribute of the JSII assemblies was extended to include
the `targets` of the depended-upon package, as well as a map of the dependencies
of that (including `targets` and `dependencies` again as relevant); and the
`externalTypes` map now contains the base types and interfaces of any
other type in the map (previously, only types exposed as part of an exported
API of the local assembly would be represented).

A `fingerprint` attribute was added, and comes in handy when trying to implement
incremental build and code generation support. Modeling it in the Assembly object
itself avoids forcing every single tool to implement it's own fingerprinting technique.

A number of redundant / antiquated attributes of the JSII assemblies were removed
as part of this change:
* `package` was redundant and always had the same value as `name`
* `typecount` was redundant (could trivially be computed from `types`)
* `nametree` duplicated information, the feature to build the data structure
  was extracted as a separate class.

The `fingerprint` in outputs of `jsii-pacmak` has been tweaked to include the version
information of `jsii-pacmak`, so code gets re-generated if the version of `jsii-pacmak`
changes. Since this makes the tests difficult, a new `--no-fingerprint` option was
added to allow opting out of the fingerprint entirely.