Make jsii-spec an interface hierarchy (#90)
The specification is usually loaded from a JSON document, so it doesn't
make sense that the types are represented as classes. This changes them
to be interfaces instead, so that it is clear no behavioral components
should be allowed.

The new `jsii-spec` package also offers a function that validates a JSII assembly
from an `any` using `jsonschema` (based on a schema generated from the
TypeScript definitions).

This makes a lot of movement happen in the "expected" assemblies from the
regression tests that are due to field re-ordering, and now the assemblies will
be alpha-sorted, so ordering is stable from now on.