pacmak: support local dependencies (and more)
Added a few features to jsii-pacmak:

* --targets: allow specifying multiple targets when running pacmak.
* --recurse: build all dependencies.
* outdir and targets default to the values from package.json/jsii
* -v: print INFO messages
* -vv: print VERBOSE messages
* --no-clean: keep generated source code after build

Outputs will go to <outdir>/<target> if multiple targets are specified
or just <outdir> if only a single target is required.

When using --recurse, --outdir applies to all modules, which means
that all the artifacts will be emitted into the same output directory.
In this case if --outdir is not specified, each target's outdir
is respected (from package.json).

Java: if one of the dependencies of a module has a local output,
it will be added as a local maven repository, so it is now possible
to depend on a local build instead of always going to maven central.

Changes to calc test library:
* Added `sphinx` output to calc test modules.
* Changed groupId to com.amazonaws.jsii