Stop hoisting external declarations in JSII assemblies (#116)
Consumers of JSII assemblies need to have all assemblies for the current package
and all it's dependencies in order to be able to reason over the entire type model for
the closure. This behavior mimics the experience one would have with C header
files, and has the side benefit of making assembly files much smaller.

This change includes the necessary update to the `jsii` compiler (so it loads transitive
dependencies, instead of only direct dependencies), and the `jsii-pacmak` (incl. the
.NET one) to correctly search for and load the dependencies.

This is a "better" fix than #113, as it doesn't cause 1+MB assemblies to be produced for
the [awslabs/aws-cdk](https://github.com/awslabs/aws-cdk).