Model license in JSII Assemblies & allow signing of Java artifacts
The License is necessary in the Maven POM in order to publish to Maven-Central. Additionally,
it is necessary that the objects are signed with a GPG key that is available on the MIT keyserver.
This change models the license as part of the JSII assemblies and requires the license to be a
valid SPDX identifier. This is translated in the Maven POM using the standard SPDX license list
(since the POM expects a full name and URL).

Additionally, POM allows signing of artifacts by setting (at least) the `gpg.keyname` property,
which can be achieved by adding `--mvn-D gpg.keyname=name@foo.bar` to the `jsii-pacmak`
invokation. The signatures are provided as `.asc` files next to the signed artifacts, as is done
by the `maven-gpg-plugin`.