Fix #128: direct dependency not found
Say a => b,c and b => c, the algorithm that traverses the dep graph has an early bail out if a dependency has been visited. c is first visited as part of the recursion into b's transitive deps, so when we reach it as a direct dependency, it is not processed, and not added to the .jsii assembly.