Output cross-language names in generated docs
Each `Target` implementation can contribute to this process by providing
two static functions that allow the sphinx generator to obtain:
* The package coordinates in the "standard" registry (e.g: maven) for an
  assembly
* The native name (or names) for a given `Type`. There can be multiple
  names in case the artifacts can be used across different languages,
  for example Javascript & TypeScript share names, but `InterfaceType`s
  are not visible in Javascript.

The C# names are not included, because the process to come up with those
names is factored in the C# code, and the dis-ambiguation is complex to
reproduce. This omission will be corrected at a later point in time.

Fixes #74