Transitive dependency issues (#129)
Fix #128: direct dependency not found

Say a => b,c and b => c, the algorithm that traverses the dep graph has an early bail out if a dependency has been visited. c is first visited as part of the recursion into b's transitive deps, so when we reach it as a direct dependency, it is not processed, and not added to the .jsii assembly.

Java: handle indirect type references

When resolving FQNs to Java, look up module within
direct and transitive deps. 

Also, we cache a map of every module referenced during
code generation, so we can account for these modules when
we generate the pom.xml file. These may include direct
dependencies but also indirect dependencies, in cases,
for example where their types are needed when generating
flat interface proxies.

Move POM.xml generation till after we generated
all the java code so we can include implicit
additional dependencies in POM (e.g. when interface
proxies reference types from a non-direct dep).

Other changes:

* Needed to add another module to the dep tree to
  simulate this situation.
* Don't build twice when using --recurse
* Add Generator.findModule which looks up a
  jsii module in all direct and transitive deps.
* Remove duplicate deps in .NET csproj
* java: don't fail build on javadocs errors
* produce tarball once and reuse for all targets
* if `outdir` is implicit, verify that .npmignore excludes it from the tarball