fix: "Malformed enum value" when using @scoped packages (#139)
Fixes #138

The code that validated the format of jsii enum values was too rigid
and didn't allow "/" in type names. When using a package from an npm
scope, the enum value will look like this `@scope/foo.EnumType/MemberName`.

Also, add a verification that the enum member exists in the enum and fail
otherwise.

Required fix in kernel, java and dotnet runtimes.