Buildable packaging
In order to be able to use Buildable's publishing
tasks, we need this repo to emit a buildable-compatible artifact.

This means that the .zip file will contain top-level directories
per target:
- java (for Maven Central)
- js (for npm)
- dotnet (for NuGet)
- ruby (for RubyGems)

Each modules that wishes to participate in the packaging process
needs to implement a "package" npm script which will produce a "dist/xxx"
directory with publishable artifacts.

Helper scripts package-xxx are available in jsii-build-tools.

Then, the pack.sh script simply merges all "dist" directories into
"pack/" and bundle.sh is unchanged (just bundles all files from pack)

* Remove +commit from version
* Rename "npm" target to "js" to conform
* Rename bundle-beta.sh to bundle.sh
* Moved the jsii-runtime webpack output to "webpack" so it won't conflate
  with "dist"

* Remove dependency between pacmak and jsii-java-runtime. It was only needed
in order to be able to build against a local version of jsii-java-runtime, so
we implement a crude heuristic for that.