Verify types of overridden methods (#115)
Verify that the type signature of overridden members is identical
to the type signature of the base class. While in many languages
return types may be narrowed in subclasses, this is not true
for (e.g.) C#.

We want to disallow this pattern in jsii so that people
will not be able to write classes that don't behave the same
in C# and other languages.

Fixes #100.