Embed dotnet-generator into pacmak (#157)
Since dotnet uses an external program to generate .NET, we need this
tool to be available when running pacmak.

This change embeds the .NET code generator (and all needed deps) into
jsii-pacmak. Size of pacmak tarball is not 4MiB (not so bad).

Also: change all modules that are not published to npm to "private: true"

Update tgz with placeholders.