Use jsii-java-runtime as a devDependency
Instead of looking up the maven repo of jsii-java-runtime in the repo
structure, define it as a devDependency and attempt to `require` it.
If it exists, it adds the local maven-repo to the user.xml maven settings
during build. Otherwise, it will just download from maven.

Also, adds a test to jsii-pacmak that recursively builds the
entire jsii-calc stack in all languages.