Make the parsing of npm pack output more robust
npm pack may have additional output with it and when we send all
of that output to path.resolve() it results in including a bunch of
other output that's not needed.

This changes makes us only take the last line of that output which should have the tarball name there.

We may want to be a little more clever and not rely on output parsing. (package name + version .tgz instead)

Fixes #172