BREAKING CHANGE: Correctly handle nested types in Java (#178)
The Java runtime had an issue where it was incorrectly attempting `Class.forName`
with a class name inferred from the jsii `fqn`. This does however not work for nested
classes as the path delimited past the point of nesting changes from `.` to `$`.

In order to address this, the `JsiiModule` class now is required to implement a
`resolveClass` method that returns the correct `Class` instance for a given `fqn`.

Moved an `enum` in `jsii-calc` from "top-level" to `class`-nested to demonstrate the
issue and verify the fix is working as intended. This rippled into changes in the
`C#` and `JS` tests for the same (since names have changed).