Correctly represent array types in the documentation
The `ref` of array types would not include the [], causing them to render in confusing
ways in the documentation. Additionally, arrays of union types would only have the [] on
the last of the options, which was incorrect. Added parentheses around the possible
element types in order to clarify the situation.