fix(jsii): Don't skip emit on TS errors when in "watch" mode
In watch mode, users expect to be able to be a little more drafty than when doing a plain compile cycle, as
they may edit code as part of a troubleshooting loop, that may involve for example leaving out unused
imports and the like. This change makes sure the `noEmitOnError` compiler option is set to false when
operating in `watch` mode, so the `.js` files get updated unconditionally.

Fixes #235