Magical Removal
Greatly simplify the implementation by removing the magic decorators and
pushing more of the work into the generated Python code. This makes the
generated Python code have more noise to seeing the structure of the API
but it means that there is less magic happening behind the scenes. This
also makes it better for tools like MyPy since this code more more like
"normal" Python code.