Don't inhereit from Interfaces
MyPy Protocols do not require inheritence to work, however we were using
that as documentation for what interfaces a particular class implements.

However, as it turns out, the Protocol class uses a metaclass, which
conflicts with the JSIIMeta in use by our JSII classes. So we'll have to
rely on the duck typing support of MyPy instead.