fix(dotnet/runtime): Incorrect callback response format
The response for a callback completion that resulted in an error should use the `err` field
name and not the `error` field name as was initially done. Additionally, the program flow
did not propagate all errors back to the `jsii-kernel`, causing false-pass on the compliance
test suite, as expectations were incorrectly set up to expect `RuntimeException` instead
of `JsiiException`.

Fixes #285