feat(jsii-reflect): library for exploring jsii type systems
jsii-reflect is a library for strong-type reflection of jsii type systems.
it creates an in-memory object graph from .jsii module specifications
and exposes a rich API for exploring the type system.

jsii-tree is a command-line utility bundled with jsii-reflect
which prints an ascii tree of a jsii type system.

oo-ascii-tree is an object-oriented library for producing ascii trees.
it supports multi-line nodes and some other cool features. 

See README files for more details.