fix(runtime): Passing 'this' to a callback from constructor
In Javascript, a constructor is allowed to pass references to `this` to
method calls (effectively passing a partially initialized instance).
When done, the kernel will assign an Object ID to `this` on the spot in
order to pass the reference through the JSII boundary. However, the
`create` API before this change would have attempted to re-allocate an
Object ID to the object (which is illegal and causes a crash).

In addition, callbacks from at least Java and .NET runtimes could not
receive JSII object references, for they would not turn them into native
objects before passing them to the implementation, resulting in a class
conversion error.