feat(jsii-spec): Model parameter optionality
Method `Parameters` now carry an `optional` flag that indicates whether
they are optional or required, and the `TypeReference#optional` field
was renamed to `TypeReference#nullable` to better reflect its semantics.

This also brings more flexibility in that it is now possible to model a
method with a nullable or defaulted argument that is followed by some
non-optional argument, and still obtain a reasonable type specification,
where previously this was an error.

Finally, in order to better reflect the type model of TypeScript and
Javascript, all `any` type references are now denoted `nullable`.

BREAKING CHANGE: JSII assemblies generated by older versions of the tool
will fail loading with this new version, and vice-versa. Re-compile your
projects in order to fix this.

Fixes #296
Fixes #414