Emphatically go to direct-level dependencies only, no need to list transitive dependencies.
The original reason for including all transitive dependencies was so
that the .NET generator could have a full view of transitively inherited
interfaces, so it can generate full proxies.

Added a test interface to verify that the build succeeds, and it does,
so apparently we don't seem to need the transitive deps anymore.