fix(java): Stop using Streams to render params
The Java cod egenerated by `jsii-pacmak` used to use `Stream`
concatenations to render parameters into an `Object` array when passing
to the low-level `jsii` engine. In the majority of cases this worked
fine, however in casesw where the last argument was an `enum`, this
caused `javac` type inferrence to impose bounds on the `Stream`'s item
type that could not be met by non-`enum` types.

Additionally, as this code demonstrates, the `Stream`-based code was
unnecessarily complicated as it is possible to initialize an array
literal in place, by using less code, which also reads simpler, and
incurs less memory churn. It's a win all over the board 🎉