feat: Register module-level stability
Adds the necessary code to detect and forward module-level stability
indications to the assembly, so it can be used by downstream languages.
This effectively removes the `markdown` attribute from the `Assembly` in
favor of using the `docs` attribute.

If a stability indication, a specific badge _must_ be present in the
accompanying `README.md` file (which as a consequence is required to
exist), to denote the maturity of the module in user-visible
documentation, too. The tool prescribes what the badge must look like.

Also, renders a `package-info.json` file for packages that have
documentation (a `README.md` file, deprecation notice, or stability
indication).