feat(jsii): Propagate stability to members
The JSII compiler now propagates an entity's own stability to it's
members that are not decorated with an explicit stability. Un-annotated
types get the stability of the assembly, un-annotated type members
(methods, properties, constructors, ...) get the stability of their
declaring type.

Note: Overridden members do not derogate! It is the responsibility of
      the overriding entity to ensure an appropriate stability is used.