fix(kernel): Correct null handling in JSON types (#523)
The JSON types were serialized "as-is" without touch, meaning if the
value was `null`, it would be forwarded as `null`, instead of `undefined`.
This caused some CDK usage patterns to crash at runtime. This changes
the behavior so that top-level `null`s in JSON types turn into `undefined`,
while `null`s nested in the JSON value will be preserved (they are valid
JSON after all!).

Additionally, tuned the Java code generator a little:
The serialization of structs creates a jackson ObjectNode to pass down
to the serialization layer, however those are turned to JSON without
processing and are not honoring the inclusion setting to ignore nulls.
Added some code in the Java generator to stop adding null values into
the ObjectNode.