feat(pacmak): support adding a suffix to Java package version (#552)
Allows setting a `jsii.targets.java.maven.versionSuffix` key that
defines a string to be appended at the end of the version number of the
Maven package. If the suffix starts with `.` or `-`, it will be appended
to the version number directly (`0.12.0` with suffix `.dp` will render
`0.12.0.dp`); but otherwise, it'll be prepended with a `-` before it is
concatenated with the version number (`0.12.0` with suffix `dp` will
render `0.12.0-dp`).