fix(jsii-diff): handle recursive types
Structural type checking would recurse endlessly if a struct had a
member with the same type as itself, or the same type as a struct that
contained itself.

Add a cache to prevent infinite recursion.

Also in this commit: `tryFindFqn` no longer explodes if the referenced
assembly couldn't be found, but simply returns `undefined` as it should.