fix(jsii): correctly inherit initializer stability (#569)
Fix the inheriting of stability tags for the initializer. It used to
inherit from the package, it now correctly inherits from the class.

When member and class stability disagree, use the one that indicates
the fewest guarantees.

Update jsii-reflect with a "load all modules" method (to save time
in a complex dependency tree) and jsii-tree with options to show
stabilities.