chore: configure dependabot (#603)
Monitor all Javascript packages as normal, and also keep an eye on the
.NET projects (JSII Runtime & related integration tests), as well as
the python and Ruby runtimes.

It is not possible to monitor the Java runtime in the same way due to
the `pom.xml` files being generated as part of the build.

Had to replace references to "private" packages with local package
references, as dependabot requires this (necessary tooling otherwise
chokes on packages that do not resolve due to not being published).