feat(jsii): emit warnings when using reserved words
Emits warnings when an exported API is named after a reserved word in
any of the supported target languages (best-effort). This warns users
that their code may have toruble compiling for certain target languages
and invites them to use a different name.

Additionally, a `--fail-on-warnings` / `--Werr` option was added to the
CLI that allows treating warnings as errors, allowing users to create
and maintain a codebase that does not rely on naming things in ways that
will cause conflicts in target languages.

Fixes #701